refactor: ♻️ refactor codebase into separate `src` directory
