+++ tsconfig.json
@@ -1,6 +1,6 @@
 {
   "compilerOptions": {
-    "baseUrl": ".",
+    "baseUrl": "src",
     "inlineSourceMap": true,
     "inlineSources": true,
     "module": "ESNext",
@@ -17,6 +17,7 @@
     ]
   },
   "include": [
-    "**/*.ts"
+    "src/*.ts",
+    "test-data/*.ts"
   ]
 }
